Fixed issue with cycling only 2 pages
It used to be that when you only had two pages, the non-visible page could only ever be on the right. So when you swiped to the left you would see nothing until the swiping was over and then the page would be put in place. Now the page is put on the correct side while you swipe.